SWIG stuff restored

+++ include/pkmn/lists.hpp
@@ -19,19 +19,19 @@
 
 namespace pkmn
 {
-    void PKMN_API get_version_list(std::vector<pkmn::pkstring> &game_vec);
+    void PKMN_API get_version_list(std::vector<pkmn::pkstring>& game_vec);
 
-    void PKMN_API get_version_group_list(std::vector<pkmn::pkstring> &game_group_vec);
+    void PKMN_API get_version_group_list(std::vector<pkmn::pkstring>& game_group_vec);
 
-    void PKMN_API get_item_list(std::vector<pkmn::pkstring> &item_vec, unsigned int game);
+    void PKMN_API get_item_list(std::vector<pkmn::pkstring>& item_vec, uint16_t game);
 
-    void PKMN_API get_pokemon_list(std::vector<pkmn::pkstring> &pokemon_vec, unsigned int game);
+    void PKMN_API get_pokemon_list(std::vector<pkmn::pkstring>& pokemon_vec, uint16_t game);
 
-    void PKMN_API get_type_list(std::vector<pkmn::pkstring> &type_vec, unsigned int gen);
+    void PKMN_API get_type_list(std::vector<pkmn::pkstring>& type_vec, uint16_t gen);
 
-    void PKMN_API get_ability_list(std::vector<pkmn::pkstring> &ability_vec, unsigned int gen);
+    void PKMN_API get_ability_list(std::vector<pkmn::pkstring>& ability_vec, uint16_t gen);
 
-    void PKMN_API get_nature_list(std::vector<pkmn::pkstring> &nature_vec);
+    void PKMN_API get_nature_list(std::vector<pkmn::pkstring>& nature_vec);
 
     void PKMN_API get_pokemon_of_type(pokemon_entry_vector_t& pkmn_vector,
                                       const pkmn::pkstring& type1,
